feat: almost finish

+++ pets/models.py
@@ -10,7 +10,7 @@ class CategoryPets(models.TextChoices):
 class Pet(models.Model):
     name = models.CharField(max_length=50)
     age = models.IntegerField()
-    weight = models.DecimalField(decimal_places=2, max_digits=5)
+    weight = models.FloatField()
     sex = models.CharField(
         max_length=20, 
         choices=CategoryPets.choices,
